moved changes from files in original branch to new files to reduce chance of merge conflicts
